services/stream_processor/requirements.txt
# FairOps Stream Processor — Dependencies
# Pinned versions from AGENT.md Section 20

apache-beam[gcp]==2.54.0
google-cloud-pubsub==2.20.0
google-cloud-bigquery==3.17.0
google-cloud-dlp==3.15.0
google-cloud-logging==3.9.0
pydantic==2.6.0
